feat: extract storage package
