+++ packages/shared/index.ts
@@ -1,3 +1,2 @@
-export * from './lib/storages';
 export * from './lib/hooks';
 export * from './lib/hoc';
