+++ packages/shared/tsup.config.ts
@@ -4,5 +4,5 @@ export default defineConfig({
   treeshake: true,
   format: ['cjs', 'esm'],
   dts: true,
-  external: ['chrome', 'react'],
+  external: ['chrome'],
 });
