+++ pages/options/src/Options.tsx
@@ -1,10 +1,6 @@
 import '@src/Options.css';
-import {
-  exampleThemeStorage,
-  useStorageSuspense,
-  withErrorBoundary,
-  withSuspense,
-} from '@chrome-extension-boilerplate/shared';
+import { useStorageSuspense, withErrorBoundary, withSuspense } from '@chrome-extension-boilerplate/shared';
+import { exampleThemeStorage } from '@chrome-extension-boilerplate/storage';
 import { ComponentPropsWithoutRef } from 'react';
 
 const Options = () => {
