+++ pages/popup/src/Popup.tsx
@@ -1,10 +1,7 @@
 import '@src/Popup.css';
-import {
-  exampleThemeStorage,
-  useStorageSuspense,
-  withErrorBoundary,
-  withSuspense,
-} from '@chrome-extension-boilerplate/shared';
+import { useStorageSuspense, withErrorBoundary, withSuspense } from '@chrome-extension-boilerplate/shared';
+import { exampleThemeStorage } from '@chrome-extension-boilerplate/storage';
+
 import { ComponentPropsWithoutRef } from 'react';
 
 const Popup = () => {
